refactor: add src folder to root, migrate server and client paths to src
