+++ __tests__/src/client/components/App.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import App from '../../../../client/src/components/App/App';
+import App from '../../../../src/client/components/App/App';
 
 test('renders a basic snapshot', () => {
   const { container } = render(<App />);
