+++ __tests__/src/client/containers/MainContainer.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react'; 
-import MainContainer from '../../../../client/src/containers/MainContainer/MainContainer';
+import MainContainer from '../../../../src/client/containers/MainContainer/MainContainer';
 
 test('renders a basic snapshot', () => {
   const { container } = render(<MainContainer />);
