+++ jest.config.ts
@@ -2,20 +2,24 @@ module.exports = {
   moduleFileExtensions: ['js', 'jsx', 'ts', 'tsx'],
   transform: {
     '^.+\\.(ts|tsx)$': 'ts-jest',
+    '^.+\\.jsx?$': 'babel-jest',
   },
   testMatch: [
     '**/__tests__/**/*.+(ts|tsx|js|jsx)',
     '**/?(*.)+(spec|test).+(ts|tsx|js|jsx)',
   ],
+  randomize: true,
+  roots: ['<rootDir>/src/', '<rootDir>/__tests__/'],
   testEnvironment: 'jsdom',
   coverageDirectory: 'coverage',
   collectCoverageFrom: ['src/**/*.{js,jsx,ts,tsx}', '!src/**/*.d.ts'],
+  coverageReporters: ['text', 'html'],
   coverageThreshold: {
     global: {
-      branches: 0,
-      functions: 0,
-      lines: 0,
-      statements: 0,
+      branches: 5,
+      functions: 5,
+      lines: 5,
+      statements: 5,
     },
   },
 };
