+++ package.json
@@ -8,7 +8,7 @@
     "test:coverage": "jest --coverage --passWithNoTests",
     "start-client": "webpack-dev-server --mode development --open --hot",
     "build-client": "webpack --mode production",
-    "start-server": "nodemon --exec ts-node -- server/server.ts",
+    "start-server": "nodemon --exec ts-node -- src/server/server.ts",
     "start": "npm-run-all start-client start-server",
     "dev": "concurrently \"npm run start-client\" \"npm run start-server\"",
     "prepare": "husky install"
@@ -47,7 +47,7 @@
   "devDependencies": {
     "@babel/core": "^7.21.8",
     "@babel/preset-env": "^7.23.2",
-    "@babel/preset-react": "^7.22.15",
+    "@babel/preset-react": "^7.23.3",
     "@testing-library/react": "^14.2.1",
     "@types/cors": "^2.8.15",
     "@types/express": "^4.17.20",
@@ -69,6 +69,7 @@
     "http-proxy-middleware": "^2.0.6",
     "husky": "^8.0.0",
     "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^3.0.3",
     "sass-loader": "^13.3.0",
