+++ src/client/scss/application.scss
