+++ src/server/controllers/FavoritesController.ts
