+++ tsconfig.json
@@ -1,9 +1,9 @@
 {
   "compilerOptions": {
-    "target": "ES2020", 
-    "module": "commonjs", 
+    "target": "ES2020",
+    "module": "commonjs",
     "lib": ["esnext", "dom"],
-    "allowJs": true, 
+    "allowJs": true,
     "outDir": "./dist",
     "rootDir": "./",
     "esModuleInterop": true,
@@ -14,16 +14,14 @@
     "noImplicitAny": true,
     "noImplicitThis": true,
     "strictNullChecks": true,
-    "resolveJsonModule": true, 
+    "resolveJsonModule": true,
     "jsx": "react-jsx"
   },
   "include": [
-    "client/**/*.ts",
-    "client/**/*.tsx",
-    "server/**/*.ts"
+    "src/**/*.ts",
+    "src/**/*.tsx",
+    "src/server/**/*.ts",
+    "src/client/index.js"
   ],
-  "exclude": [
-    "node_modules",
-    "dist"
-  ]
+  "exclude": ["node_modules", "dist"]
 }
